컨트롤러 분리, 챌린지 정산 추가
