+++ src/main/java/com/sist/haebollangce/config/ServletConfig.java
@@ -4,6 +4,7 @@ import com.sist.haebollangce.user.util.AES256;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.web.multipart.commons.CommonsMultipartResolver;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
@@ -18,6 +19,7 @@ import java.io.UnsupportedEncodingException;
 @Configuration
 @EnableWebMvc
 @ComponentScan(basePackages = "com.sist")
+@EnableScheduling // 스케줄러 사용 어노테이션 추가 - jaesik
 public class ServletConfig implements WebMvcConfigurer {
 
     private static final String[] CLASSPATH_RESOURCE_LOCATIONS = {"classpath:/resources/", "classpath:/static/" };
